Backend de Spring-Boot POST,PUT,UPDATE,DELETE

+++ src/main/java/com/tutorial/cloudinaryrest/repository/ImagenRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ImagenRepository extends JpaRepository<Imagen, Integer> {
+public interface ImagenRepository extends JpaRepository<Imagen, Long> {
     List<Imagen> findByOrderById();
 }
